Removed obsolete malloc.h and memalign() in favour of standard `std::aligned_alloc`
This fixes a compiler error on OpenBSD, which does not have malloc.h.

On other UNIXes the header is mostly obsolete. It was only checked here to detect the old memalign() function which was replaced by posix_memalign() many many years ago.

And at this point, even posix_memalign() is not needed because we now require C++17 which has std::aligned_alloc.